refactor: InputProps type을 common에서 mypage로 위치 이동
