--- src/models/Common.ts
@@ -1,9 +1,0 @@
-export interface InputProps {
-  type: string;
-  id: string;
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  placeholder: string;
-  error?: boolean;
-  message?: string;
-}
