+++ src/models/MyPage.ts
@@ -14,6 +14,16 @@ export interface ProfileDetailsProps {
   detail: string;
 }
 
+export interface InputProps {
+  type: string;
+  id: string;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  placeholder: string;
+  error?: boolean;
+  message?: string;
+}
+
 export interface ThemeSelectionProps {
   isDarkMode: boolean;
   setIsDarkMode: (theme: boolean) => void;
